+++ bin/tester/index.js
@@ -1,3 +1,6 @@
+import fetch from 'node-fetch';
+import { pathExists } from 'fs-extra';
+import path from 'path';
 import denodeify from 'denodeify';
 import { confirm } from 'node-ask';
 import setupDebug from 'debug';
@@ -85,14 +88,14 @@ async function prepareAppOrBuild({
 
       const child = await startApp({
         scriptName: buildScriptName,
-        // Make storybook build as quiet as possible
+        // Make Storybook build as quiet as possible
         args: [
           '--',
           '-o',
           buildDirName,
           ...(storybookVersion &&
           gte(storybookVersion, STORYBOOK_CLI_FLAGS_BY_VERSION['--loglevel'])
-            ? ['--loglevel', 'error']
+            ? ['--loglevel', log.level === 'verbose' ? 'verbose' : 'error']
             : []),
         ],
         inheritStdio: true,
@@ -102,12 +105,24 @@ async function prepareAppOrBuild({
       await new Promise((res, rej) => {
         child.on('error', rej);
         child.on('close', code => {
-          if (code > 0) {
+          if (code !== 0) {
             rej(new Error(`${buildScriptName} script exited with code ${code}`));
           }
           res();
         });
       });
+    }
+
+    const exists = await pathExists(path.join(buildDirName, 'iframe.html'));
+
+    if (!exists) {
+      if (buildScriptName) {
+        throw new Error(`Storybook was not build succesfully, there are likely errors above`);
+      } else {
+        throw new Error(dedent`
+          It looks like your Storybook build (to directory: ${buildDirName}) failed, as that directory is empty. Perhaps something failed above?
+        `);
+      }
     }
 
     log.info(dedent`Uploading your built Storybook...`);
@@ -243,6 +258,7 @@ export async function runTest({
   createTunnel = true,
   originalArgv = false,
   sessionId,
+  allowConsoleErrors,
 }) {
   const names = getProductVariables();
 
@@ -318,13 +334,21 @@ export async function runTest({
 
   const { version: storybookVersion, viewLayer, addons } = await getStorybookInfo();
   debug(
-    `Detected package version: ${packageVersion}, storybook version: ${storybookVersion}, view layer: ${viewLayer}, addons: ${
+    `Detected package version: ${packageVersion}, Storybook version: ${storybookVersion}, view layer: ${viewLayer}, addons: ${
       addons.length ? addons.map(addon => addon.name).join(', ') : 'none'
     }`
   );
 
   let exitCode = 5;
+  let errorCount = 0;
+  let specCount = 0;
+  let componentCount = 0;
+  let changeCount = 0;
+  let buildNumber = 0;
+  let snapshotCount = 0;
   let exitUrl = '';
+  let diffs;
+  let buildStatus;
 
   const { cleanup, isolatorUrl, cachedUrl } = await prepareAppOrBuild({
     storybookVersion,
@@ -342,6 +366,18 @@ export async function runTest({
 
   debug(`Connecting to ${isolatorUrl} (cachedUrl ${cachedUrl})`);
 
+  if (
+    await fetch(isolatorUrl)
+      .then(_ => true)
+      .catch(e => {
+        throw new Error(
+          `Storybook was not build succesfully, or provided url (${isolatorUrl}) wasn't reachable, there are likely errors above`
+        );
+      })
+  ) {
+    debug(`connected to ${isolatorUrl} success`);
+  }
+
   log.info(
     `Uploading and verifying build (this may take a few minutes depending on your connection)`
   );
@@ -352,17 +388,18 @@ export async function runTest({
       list,
       isolatorUrl,
       verbose,
+      allowConsoleErrors,
     });
 
     const environment = await getEnvironment();
 
-    const {
+    ({
       createBuild: {
-        number,
+        number: buildNumber,
         snapshotCount,
         specCount,
         componentCount,
-        webUrl,
+        webUrl: exitUrl,
         app: {
           account: {
             features: { diffs },
@@ -390,13 +427,11 @@ export async function runTest({
         viewLayer,
       },
       isolatorUrl,
-    });
+    }));
 
-    exitUrl = webUrl;
-
-    const onlineHint = `View it online at ${webUrl}`;
+    const onlineHint = `View it online at ${exitUrl}`;
     log.info(dedent`
-      Started Build ${number} (${pluralize(componentCount, 'component')}, ${pluralize(
+      Started Build ${buildNumber} (${pluralize(componentCount, 'component')}, ${pluralize(
       specCount,
       'story'
     )}, ${pluralize(snapshotCount, 'snapshot')}).
@@ -406,19 +441,16 @@ export async function runTest({
 
     if (doExitOnceSentToChromatic) return { exitCode: 0, exitUrl };
 
-    const {
-      status,
-      autoAcceptChanges: buildAutoAcceptChanges, // if it is the first build, this may have been set
-      changeCount,
-      errorCount,
-    } = await waitForBuild(client, { buildNumber: number }, { diffs });
+    const buildOutput = await waitForBuild(client, { buildNumber }, { diffs });
 
-    switch (status) {
+    ({ changeCount, errorCount, status: buildStatus } = buildOutput);
+
+    switch (buildStatus) {
       case 'BUILD_PASSED':
         log.info(
           diffs
-            ? `Build ${number} passed! ${onlineHint}.`
-            : `Build ${number} published! ${onlineHint}.`
+            ? `Build ${buildNumber} passed! ${onlineHint}.`
+            : `Build ${buildNumber} published! ${onlineHint}.`
         );
         exitCode = 0;
         break;
@@ -427,12 +459,12 @@ export async function runTest({
       case 'BUILD_PENDING':
       case 'BUILD_DENIED':
         log.info(dedent`
-          Build ${number} has ${pluralize(changeCount, 'change')}.
+          Build ${buildNumber} has ${pluralize(changeCount, 'change')}.
 
           ${onlineHint}.
         `);
         console.log('');
-        exitCode = doExitZeroOnChanges || buildAutoAcceptChanges ? 0 : 1;
+        exitCode = doExitZeroOnChanges || buildOutput.autoAcceptChanges ? 0 : 1;
         if (exitCode !== 0) {
           log.info(dedent`
             Pass --exit-zero-on-changes if you want this command to exit successfully in this case.
@@ -445,12 +477,12 @@ export async function runTest({
         log.info(
           diffs
             ? dedent`
-                Build ${number} has ${pluralize(errorCount, 'error')}.
+                Build ${buildNumber} has ${pluralize(errorCount, 'error')}.
               
                 ${onlineHint}.
               `
             : dedent`
-                Build ${number} was published but we found errors.
+                Build ${buildNumber} was published but we found errors.
               
                 ${onlineHint}.
               `
@@ -460,12 +492,12 @@ export async function runTest({
       case 'BUILD_TIMED_OUT':
       case 'BUILD_ERROR':
         log.info(dedent`
-          Build ${number} has failed to run. Our apologies. Please try again.
+          Build ${buildNumber} has failed to run. Our apologies. Please try again.
         `);
         exitCode = 3;
         break;
       default:
-        throw new Error(`Unexpected build status: ${status}`);
+        throw new Error(`Unexpected build status: ${buildStatus}`);
     }
   } catch (e) {
     if (
@@ -522,5 +554,5 @@ export async function runTest({
     }
   }
 
-  return { exitCode, exitUrl };
+  return { exitCode, exitUrl, buildNumber, errorCount, changeCount, specCount, componentCount };
 }